+++ packages/rocketh-export/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-export
 
+## 0.6.0
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.6.0
+
 ## 0.5.21
 
 ### Patch Changes